refactor: reorganize admin components into components folder
