+++ apps/guardian-ui/src/components/BitcoinNodeCard.tsx
